Clase14S-Aviones-mao/Aviones/src/main/com/dh/aviones/dao/impl/AvionDAOH2.java
package main.com.dh.aviones.dao.impl;

import main.com.dh.aviones.dao.IDao;
import main.com.dh.aviones.model.Avion;
import org.apache.log4j.Logger;

import java.sql.*;
import java.time.LocalDate;
import java.util.ArrayList;
import java.util.List;

public class AvionDAOH2 implements IDao<Avion> {


    //Configuraciones de la conexión a la base de datos
    private final static String DB_JDBC_DRIVER = "org.h2.Driver";
    private final static String DB_URL = "jdbc:h2:~/db_aviones";
    private final static String DB_USER = "sa";
    private final static String DB_PASSWORD = "";
    private final static Logger logger = Logger.getLogger(AvionDAOH2.class);


    @Override
    public Avion guardar(Avion avion) {
        Connection connection = null;
        PreparedStatement preparedStatement = null;

        try {
            Class.forName(DB_JDBC_DRIVER);
            connection = DriverManager.getConnection(DB_URL, DB_USER, DB_PASSWORD);
            preparedStatement = connection.prepareStatement("INSERT INTO avion VALUES(?, ?, ?, ?, ?)");
            preparedStatement.setLong(1,avion.getId());
            preparedStatement.setString(2,avion.getMarca());
            preparedStatement.setString(3, avion.getModelo());
            preparedStatement.setString(4, avion.getMatricula());
            preparedStatement.setObject(5, avion.getFechaEntrada());

            preparedStatement.executeUpdate();
            preparedStatement.close();

        } catch (ClassNotFoundException e) {
            e.printStackTrace();
        } catch (SQLException throwables) {
            throwables.printStackTrace();
        }

        return avion;
    }

    @Override
    public void eliminar(Long id) {
        Connection connection = null;
        PreparedStatement preparedStatement = null;

        try {
            Class.forName(DB_JDBC_DRIVER);
            connection = DriverManager.getConnection(DB_URL, DB_USER, DB_PASSWORD);
            preparedStatement = connection.prepareStatement("DELETE FROM avion WHERE id=?");
            preparedStatement.setLong(1,id);

            preparedStatement.executeUpdate();
            preparedStatement.close();

        } catch (ClassNotFoundException e) {
            e.printStackTrace();
        } catch (SQLException throwables) {
            throwables.printStackTrace();
        }


    }

    @Override
    public Avion buscar(Long id) {
        Connection connection = null;
        PreparedStatement preparedStatement = null;
        Avion avion = null;

        try {
            Class.forName(DB_JDBC_DRIVER);
            connection = DriverManager.getConnection(DB_URL, DB_USER, DB_PASSWORD);
            preparedStatement = connection.prepareStatement("SELECT * FROM avion WHERE id=?");
            preparedStatement.setLong(1,id);

            ResultSet result = preparedStatement.executeQuery();

            while(result.next()) { //Mientras haya resultados
                Long idAvion = result.getLong("id");
                String marca = result.getString("marca");
                String modelo = result.getString("modelo");
                String matricula = result.getString("matricula");
                LocalDate fechaEntrada = result.getDate("fechaEntrada").toLocalDate();

                avion = new Avion(idAvion, marca, modelo, matricula, fechaEntrada);
            }

            preparedStatement.close();

        } catch (ClassNotFoundException e) {
            e.printStackTrace();
        } catch (SQLException throwables) {
            throwables.printStackTrace();
        }

        return avion;
    }

    @Override
    public List buscartodos() {
        Connection connection = null;
        PreparedStatement preparedStatement = null;
        List<Avion> aviones = new ArrayList<Avion>();

        try {
            Class.forName(DB_JDBC_DRIVER);
            connection = DriverManager.getConnection(DB_URL, DB_USER, DB_PASSWORD);
            preparedStatement = connection.prepareStatement("SELECT * FROM avion");

            ResultSet result = preparedStatement.executeQuery();

            while(result.next()) {
                Long idAvion = result.getLong("id");
                String marca = result.getString("marca");
                String modelo = result.getString("modelo");
                String matricula = result.getString("matricula");
                LocalDate fechaEntrada = result.getDate("fechaEntrada").toLocalDate();

                Avion avion = new Avion(idAvion, marca, modelo, matricula, fechaEntrada);
                aviones.add(avion);
            }
            preparedStatement.close();

        } catch (ClassNotFoundException e) {
            e.printStackTrace();
        } catch (SQLException throwables) {
            throwables.printStackTrace();
        }

        return aviones;
    }
}
